structure improvement

+++ diylc/diylc-core/src/main/java/org/diylc/plugins/chatbot/model/AiPoint.java
@@ -1,4 +1,4 @@
 package org.diylc.plugins.chatbot.model;
 
-public record AiPoint(double x, double y) {
+public record AiPoint(int x, int y) {
 }
